# mobile/openapi/mobile-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KarActionCreateRequest": {
        "description": "What's sent to log one kar movement. The team and timestamp are filled\nin by the server, never by the client.",
        "properties": {
          "kar_id": {
            "title": "Kar Id",
            "type": "integer"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "status_id": {
            "title": "Status Id",
            "type": "integer"
          }
        },
        "required": [
          "kar_id",
          "status_id",
          "latitude",
          "longitude"
        ],
        "title": "KarActionCreateRequest",
        "type": "object"
      },
...
      "KarActionResponse": {
        "description": "One logged kar movement, with the names needed for the history table.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kar_id": {
            "title": "Kar Id",
            "type": "integer"
          },
          "kar_nummer": {
            "title": "Kar Nummer",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "status_id": {
            "title": "Status Id",
            "type": "integer"
          },
          "status_name": {
            "title": "Status Name",
            "type": "string"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          }
        },
        "required": [
          "id",
          "kar_id",
          "kar_nummer",
          "status_id",
          "status_name",
          "team_id",
          "team_name",
          "latitude",
          "longitude",
          "recorded_at",
          "user_name"
        ],
        "title": "KarActionResponse",
        "type": "object"
      },
...
      "KarMapResponse": {
        "description": "The full Kar Map payload: one array per layer. Rows with no\nlatitude/longitude are included too \u2014 the frontend excludes them from\nthe map itself but still lists them in the side panel.",
        "properties": {
          "afleverlocaties": {
            "items": {
              "$ref": "#/components/schemas/KarMapAfleverlocatieRow"
            },
            "title": "Afleverlocaties",
            "type": "array"
          },
          "distributiepunten": {
            "items": {
              "$ref": "#/components/schemas/KarMapDistributiepuntRow"
            },
            "title": "Distributiepunten",
            "type": "array"
          },
          "karren": {
            "items": {
              "$ref": "#/components/schemas/KarMapKarRow"
            },
            "title": "Karren",
            "type": "array"
          }
        },
        "required": [
          "karren",
          "afleverlocaties",
          "distributiepunten"
        ],
        "title": "KarMapResponse",
        "type": "object"
      },
      "KarPlanningFestivalResponse": {
        "description": "One festival column of the Kar Planning report.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "KarPlanningFestivalResponse",
        "type": "object"
      },
      "KarPlanningReportResponse": {
        "description": "The whole Kar Planning report: the active festivals of the requested\nseason (one extra table column each, in start-date order) plus one row\nper kar. `festivals` is empty when no season was requested.",
        "properties": {
          "festivals": {
            "items": {
              "$ref": "#/components/schemas/KarPlanningFestivalResponse"
            },
            "title": "Festivals",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/KarPlanningResponse"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "festivals",
          "rows"
        ],
        "title": "KarPlanningReportResponse",
        "type": "object"
      },
...
      "KarTrackerGroundplanResponse": {
        "description": "One ground-plan overlay: its name and south-west/north-east corner\ncoordinates. The image itself is served separately by\nGET /groundplans/{id}/image; updated_at lets the frontend cache-bust\nthat URL after the image is replaced.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ne_latitude": {
            "title": "Ne Latitude",
            "type": "number"
          },
          "ne_longitude": {
            "title": "Ne Longitude",
            "type": "number"
          },
          "sw_latitude": {
            "title": "Sw Latitude",
            "type": "number"
          },
          "sw_longitude": {
            "title": "Sw Longitude",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "sw_latitude",
          "sw_longitude",
          "ne_latitude",
          "ne_longitude",
          "updated_at"
        ],
        "title": "KarTrackerGroundplanResponse",
        "type": "object"
      },
...
      "MobileKarScanResponse": {
        "description": "The kar found for a scanned QR code, with its latest movements.",
        "properties": {
          "kar": {
            "$ref": "#/components/schemas/KarActionKarOption"
          },
          "recent_actions": {
            "items": {
              "$ref": "#/components/schemas/KarActionResponse"
            },
            "title": "Recent Actions",
            "type": "array"
          }
        },
        "required": [
          "kar",
          "recent_actions"
        ],
        "title": "MobileKarScanResponse",
        "type": "object"
      },
      "MobileModule2PermissionsResponse": {
        "description": "What the calling user may do in the phone's KarTracker: KarScan (the\nweb's \"Manuele kar beweging\"), Kar Planning and Kar Map. There is no\ndelete flag: deleting a movement is web-only.",
        "properties": {
          "can_create_actions": {
            "title": "Can Create Actions",
            "type": "boolean"
          },
          "can_view_actions": {
            "title": "Can View Actions",
            "type": "boolean"
          },
          "can_view_karmap": {
            "title": "Can View Karmap",
            "type": "boolean"
          },
          "can_view_karplanning": {
            "title": "Can View Karplanning",
            "type": "boolean"
          }
        },
        "required": [
          "can_view_actions",
          "can_create_actions",
          "can_view_karplanning",
          "can_view_karmap"
        ],
        "title": "MobileModule2PermissionsResponse",
        "type": "object"
      },
...
      "MobileModule3PermissionsResponse": {
        "description": "What the calling user may do on the phone's Intervention Requests\n\"Akties\" screen, so the app only shows buttons the web roles allow.\n(There is no delete flag: deleting is web-only.)",
        "properties": {
          "can_create_requests": {
            "title": "Can Create Requests",
            "type": "boolean"
          },
          "can_edit_requests": {
            "title": "Can Edit Requests",
            "type": "boolean"
          },
          "can_view_requests": {
            "title": "Can View Requests",
            "type": "boolean"
          }
        },
        "required": [
          "can_view_requests",
          "can_create_requests",
          "can_edit_requests"
        ],
        "title": "MobileModule3PermissionsResponse",
        "type": "object"
      },
...
      "MobileSeasonResponse": {
        "description": "One open season, for the phone's Kar Planning season dropdown.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "MobileSeasonResponse",
        "type": "object"
      },
...
      "PlanKarOption": {
        "description": "One entry of a \"Plan a kar\" dropdown (a team or a delivery location).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "PlanKarOption",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RW Crew mobile API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mobile/v1/module-2/groundplans": {
      "get": {
        "description": "The ground plans overlaid on the Kar Map: name and corner coordinates.",
        "operationId": "get_groundplans_api_mobile_v1_module_2_groundplans_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KarTrackerGroundplanResponse"
                  },
                  "title": "Response Get Groundplans Api Mobile V1 Module 2 Groundplans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Groundplans",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/groundplans/{groundplan_id}/image": {
      "get": {
        "description": "One ground plan's image bytes (PNG or JPEG).",
        "operationId": "get_groundplan_image_api_mobile_v1_module_2_groundplans__groundplan_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "groundplan_id",
            "required": true,
            "schema": {
              "title": "Groundplan Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Groundplan Image",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/kar-actions": {
      "post": {
        "description": "Log one kar movement, exactly like the web's Save button: the team and\ntimestamp are set here, and the kar's latest status/location follow.",
        "operationId": "create_kar_action_api_mobile_v1_module_2_kar_actions_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KarActionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KarActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Kar Action",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/kar-map": {
      "get": {
        "description": "Every kar, afleverlocatie and distributiepunt for the map, identical\nto the web's Kar Map data (rows without coordinates included).",
        "operationId": "get_kar_map_api_mobile_v1_module_2_kar_map_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KarMapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Kar Map",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/kar-planning": {
      "get": {
        "description": "The Kar Planning report, identical to the web's; with a season, one\nplanned afleverlocatie per active festival of it.",
        "operationId": "get_kar_planning_api_mobile_v1_module_2_kar_planning_get",
        "parameters": [
          {
            "in": "query",
            "name": "season_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season Id"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KarPlanningReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Kar Planning",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/kar-statuses": {
      "get": {
        "description": "The status dropdown, gated by the movement screen itself (like the\nweb's /kar-actions/lookups) so no KarStatussen rights are needed.",
        "operationId": "list_kar_statuses_api_mobile_v1_module_2_kar_statuses_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanKarOption"
                  },
                  "title": "Response List Kar Statuses Api Mobile V1 Module 2 Kar Statuses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Kar Statuses",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/karren/by-nummer/{kar_nummer}": {
      "get": {
        "description": "The kar whose QR code was scanned, plus its last few movements.\n\nThe scanned text is trimmed and matched case-insensitively, so a QR code\nprinted as \"b001 \" still finds kar \"B001\".",
        "operationId": "get_kar_by_nummer_api_mobile_v1_module_2_karren_by_nummer__kar_nummer__get",
        "parameters": [
          {
            "in": "path",
            "name": "kar_nummer",
            "required": true,
            "schema": {
              "title": "Kar Nummer",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileKarScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Kar By Nummer",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/permissions": {
      "get": {
        "description": "What this user may do in the phone's KarTracker, so the app only\nshows the menu items and buttons the web roles allow.",
        "operationId": "get_my_permissions_api_mobile_v1_module_2_permissions_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileModule2PermissionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Permissions",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
    "/api/mobile/v1/module-2/seasons": {
      "get": {
        "description": "The open seasons (the same ones the web header's season selector\noffers), newest first, for the Kar Planning season dropdown.",
        "operationId": "list_open_seasons_api_mobile_v1_module_2_seasons_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MobileSeasonResponse"
                  },
                  "title": "Response List Open Seasons Api Mobile V1 Module 2 Seasons Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Open Seasons",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
...
    "/api/mobile/v1/module-3/permissions": {
      "get": {
        "description": "What this user may do on the requests screen (view / create / edit).",
        "operationId": "get_my_permissions_api_mobile_v1_module_3_permissions_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileModule3PermissionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Permissions",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
...
  }
}